backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic-settings==2.2.1
redis==5.0.4
prometheus-client==0.20.0
# aioredis shim conflicts on Python 3.11 (duplicate TimeoutError); use redis asyncio API
SQLAlchemy[asyncio]==2.0.30
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.13.1
pytest==8.2.2
pytest-asyncio==0.23.7
respx==0.21.1
psutil==5.9.8
nvidia-ml-py3==7.352.0
docker==6.1.3
# Pin requests for docker SDK compatibility (avoids "Not supported URL scheme http+docker")
requests==2.31.0
requests-unixsocket==0.3.0